refactor: Move ThermocyclerMovementFlagger to its own file.
